5311 - DocketEntry Entity - set servedPartiesCode when setting doc as served
