+++ web-client/src/applicationContext.js
@@ -5,6 +5,7 @@ import {
 } from '../../shared/src/business/entities/cases/Case';
 import {
   DocketEntry,
+  getServedPartiesCode,
   isServed,
 } from '../../shared/src/business/entities/DocketEntry';
 import { ErrorFactory } from './presenter/errors/ErrorFactory';
@@ -116,7 +117,6 @@ import {
   formatDocketEntry,
   getFilingsAndProceedings,
   getFormattedCaseDetail,
-  getServedPartiesCode,
   sortDocketEntries,
 } from '../../shared/src/business/utilities/getFormattedCaseDetail';
 import { forwardMessageInteractor } from '../../shared/src/proxies/messages/forwardMessageProxy';
